+++ tests/conftest.py
@@ -10,12 +10,17 @@ from ember_mug.connection import EmberMugConnection
 
 @pytest.fixture
 def ember_mug():
-    yield EmberMug(BLEDevice(address='32:36:a5:be:88:cb', name='Ember Ceramic Mug'))
+    yield EmberMug(
+        BLEDevice(
+            address='32:36:a5:be:88:cb',
+            name='Ember Ceramic Mug',
+        )
+    )
 
 
 @pytest_asyncio.fixture
 async def mug_connection():
     mug = EmberMug(BLEDevice(address='32:36:a5:be:88:cb', name='Ember Ceramic Mug'))
     connection = EmberMugConnection(mug)
-    connection.client = AsyncMock()
+    connection._client = AsyncMock()
     yield connection